fix(clearing): put the position book and the rolling window in fixed point
LeveragedPosition carried quantity, entry price, equity and the contract
multiplier as doubles, and Account carried equity, marks and the 30-day
rolling notional the same way. Margin and liquidation are the last place in
the engine that should be computing in double: a double quantity cannot even
represent the fills that produced it, so every notional, maintenance-margin
requirement and liquidation price derived from one starts from a value that is
already approximate, by an amount that depends on the build.

The rolling counter showed what that costs. It was accumulated with += and -=,
so a five-billion fill swallowed every 1e-8 fill recorded beside it; when the
large one aged out, the subtraction took the total negative and the clamp at
the end of evictExpired rewrote it as zero, and the five hundred fills still
inside the window were gone. The clamp is removed rather than kept, because
the sum is now exact: what eviction subtracts is exactly what recordFill
added, and a negative total would mean a negative fill was recorded, not that
the sum drifted.

Every aggregate goes through one pair of per-leg helpers, so the notional and
the unrealised PnL a margin check reads cannot disagree with what a close
books. The double-taking overloads stay on the input side and quantise once,
at the boundary, for the C ABI, the bindings and configuration; the accessors
return Volume / Price. The liquidation engine's own ranking and partial-fill
arithmetic is unchanged and reads the legs through five conversions spelled
out at the top of the file.

+++ tests/test_cross_cascade.cpp
@@ -78,12 +78,12 @@ TEST(CrossCascade, CascadeMarksAreSymbolLocal)
   // Set ETH mark explicitly to a known value; cascade must NOT
   // touch it.
   a.setMark(ETH, 2'900.0);
-  const double ethMarkBefore = a.markFor(ETH);
+  const double ethMarkBefore = a.markFor(ETH).toDouble();
 
   (void)e.onMark(BTC, 45'000.0);
 
   // ETH mark unchanged by the BTC cascade — symbol-local design.
-  EXPECT_DOUBLE_EQ(a.markFor(ETH), ethMarkBefore);
+  EXPECT_DOUBLE_EQ(a.markFor(ETH).toDouble(), ethMarkBefore);
 }
 
 TEST(CrossCascade, CascadeRunsAfterCrossMarginLiquidationFires)
